+++ packages/uipath_langchain_client/src/uipath_langchain_client/utils.py
@@ -13,10 +13,6 @@ from uipath.llm_client.utils.exceptions import (
     UiPathTooManyRequestsError,
     UiPathUnprocessableEntityError,
 )
-from uipath.llm_client.utils.model_family import (
-    ANTHROPIC_MODEL_NAME_KEYWORDS,
-    is_anthropic_model_name,
-)
 from uipath.llm_client.utils.retry import RetryConfig
 
 __all__ = [
@@ -34,6 +30,4 @@ __all__ = [
     "UiPathServiceUnavailableError",
     "UiPathGatewayTimeoutError",
     "UiPathTooManyRequestsError",
-    "ANTHROPIC_MODEL_NAME_KEYWORDS",
-    "is_anthropic_model_name",
 ]